requirements/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.2.2
djangorestframework-api-key==2.3.0
django-filter==23.5
psycopg[binary]
psycopg2-binary
dj-database-url==2.1.0
requests==2.31.0
drf-yasg==1.21.5
qrcode==7.4.2
gunicorn
python-decouple
python-dateutil
django-money
brazilnum
Pillow==9.4.0
pika
Celery
django-celery-beat
